+++ tests/test_etheroll.py
@@ -284,7 +284,8 @@ class TestEtheroll(unittest.TestCase):
                 'input': (
                         '0xdc6dd152000000000000000000000000000'
                         '000000000000000000000000000000000000e'),
-                'timeStamp': '0x5ac80f92',
+                # also note how the timestamp is not in hexadecimal, refs #6
+                'timeStamp': '1523060626',
                 'to': '0x048717ea892f23fb0126f00640e2b18072efd9d2',
                 'value': '500000000000000000',
             },
@@ -300,7 +301,7 @@ class TestEtheroll(unittest.TestCase):
                 'input': (
                         '0xdc6dd152000000000000000000000000000'
                         '0000000000000000000000000000000000002'),
-                'timeStamp': '0x5ac80f0e',
+                'timeStamp': '1523060494',
                 'to': '0x048717ea892f23fb0126f00640e2b18072efd9d2',
                 'value': '450000000000000000',
             },
@@ -322,7 +323,7 @@ class TestEtheroll(unittest.TestCase):
                     'bet_size_ether': 0.5,
                     'roll_under': 14,
                     'block_number': '5394094',
-                    'timestamp': '0x5ac80f92',
+                    'timestamp': '1523060626',
                     'datetime': datetime(2018, 4, 7, 0, 23, 46),
                     'transaction_hash': (
                         '0x0440f1013a5eafd88f16be6b5612b6e'
@@ -332,7 +333,7 @@ class TestEtheroll(unittest.TestCase):
                     'bet_size_ether': 0.45,
                     'roll_under': 2,
                     'block_number': '5394085',
-                    'timestamp': '0x5ac80f0e',
+                    'timestamp': '1523060494',
                     'datetime': datetime(2018, 4, 7, 0, 21, 34),
                     'transaction_hash': (
                         '0x72def66d60ecc85268c714e71929953'