+++ tests/test_utils.py
@@ -1,6 +1,7 @@
 import unittest
+from datetime import datetime
 
-from pyetheroll.utils import EtherollUtils
+from pyetheroll.utils import EtherollUtils, timestamp2datetime
 
 
 class TestEtherollUtils(unittest.TestCase):
@@ -15,3 +16,16 @@ class TestEtherollUtils(unittest.TestCase):
         chances_win = 100
         payout = EtherollUtils.compute_profit(bet_size, chances_win)
         self.assertEqual(payout, None)
+
+
+class TestUtils(unittest.TestCase):
+
+    def test_timestamp2datetime(self):
+        self.assertEqual(
+            timestamp2datetime('1566645978'),
+            datetime(2019, 8, 24, 11, 26, 18)
+        )
+        self.assertEqual(
+            timestamp2datetime('0x5d611eda'),
+            datetime(2019, 8, 24, 11, 26, 18)
+        )